arm64: dts: marvell: prepare 'cpus' node of armada-ap810 for multi-AP
This commit prepares the 'cpus' of armada-ap810.dtsi to support
including armada-ap810.dtsi multiple times, to describe several APs in
the same SoC.

To this effect, a macro called CPU_NUM() is introduced to generate the
cpu nodes unit addresses, while CPU_NUM_REG() is introduced to
generate the 'reg' property value.

The AP number is encoded in the upper 16 bits, and the CPU number in
the lower 16 bits. We simply use text concatenation to achieve this,
which is why CPU numbers are all described with 4 digits.

The AP810_ID define is intended to be defined by the file including
armada-ap810.dtsi.

In addition to this, the "numa-node-id" property is added, using the
AP number. Indeed, each AP has its own DDR controller and its own
attached memory. All APs can access the memory of other APs, but it
works like a NUMA system with different memory locality
properties. Therefore, it makes sense to describe how the CPUs are
grouped together as NUMA nodes.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>