arm64: dts: marvell: document armada-cp110 expected definitions
The armada-cp110.dtsi file is included several times to describe
multiple instances of the CP110 in a given SoC. This is done by
defining a number of C pre-processor macros before including this
file.

This commit adds a comment at the top of the armada-cp110.dtsi to
document which macros we expected to be available because this file
gets included.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>