arm64: dts: marvell: differentiate bus/cpu PCIe addresses in armada-cp110.dtsi
On Armada 7K/8K the bus and CPU addresses for PCIe MEM and IO regions
were identical, so we had only one set of macros to define them.

On Armada 8KP, they are going to be different, so let's separate them
in two different macros. Of course, for now their value is still
identical, but future commits introducing Armada 8KP support will
change that.

On Armada 7K/8K, this commit is a no-op.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>