pinctrl: mvebu: add driver for Marvell AP810 pin-muxing
The Marvell AP810 is one of the hardware block part of the Marvell
Armada 8KP SoCs. It contains a pin-muxing controller, which is
supported by the driver proposed in this commit.

It is similar to the driver used for the AP806, just with different
pin functionalities, and a different number of pins. But the
differences in pin functionalities were sufficient that having a
common driver would make the result a lot less readable than two
separate and simple drivers.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>