pinctrl: mvebu/cp110: call mvebu_pinctrl_assign_variant() only once
In armada_cp110_pinctrl_probe(), we assign the ->variant field of
mvebu_mpp_mode depending on the pin number, by calling
mvebu_pinctrl_assign_variant().

However, as we are going to add more variants, having this call in
every case of the switch will become a bit annoying, so let's
factorize this into a single mvebu_pinctrl_assign_variant() call.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>