net: mvpp2: make multi queue mode the default mode
The PPV2.2 controller has a multi_queue mode, that allows to have
multiple queues per port. This mode is needed to have RSS available, and
offers some nice overall performance improvements. This mode has been
usable through the use of a module parameter, this commit makes it the
default value.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>