net: mvpp2: Add vlan prio based flow steering
The PPv2 controller has a powerfull but complex classifier, that uses
numerous TCAMs and mach engines to perform classification of ingress
traffic.

This commit adds a first implementation of flow steering, limited for
now to matching incoming traffic based on the 802.3Q priority.

This is done using the QoS tables, that are included in the C2 TCAM
match engine. We don't actually use the TCAM to match the key in our
case, since the controller has a built-in mecanism to assign the
destination rxq based on the priority extracted by the Header Parser.

This interfaces with the ethtool rxfnc infrastructure.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>