net: mvpp2: avoid entering an interrupt loop
When the link interrupt is used alongside phylink (when a port does not
have a PHY) and if this port is configured as a fixed-link, an IRQ loop
can be seen when disconnecting the SFP cable. This is due to the
mac_config() function disabling and enabling the port as port of its
reconfiguration, which triggers an interrupt. As the link interrupt
handler, and the Phylink mac_config() function are called
asynchronously, an IRQ loop happens.

The solution is to mask the interrupts while reconfiguring a port, and
to acknowledge the interrupt before unmasking them. To avoid missing a
link status interrupt, special care is done in mac_config() to detect if
the link status has changed in the meantime.

Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>